pyproject.toml
[tool.poetry]
name = "api-backend"
version = "0.1.0"
description = ""
authors = ["MrLoydHD <hf.correia@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
botocore = "^1.35.44"
python-jose = "^3.3.0"
boto3 = "^1.35.44"
cryptography = "^43.0.3"
httpx = "^0.27.2"
pymysql = "^1.1.1"
pytest = "^8.3.3"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
tox = "^4.21.2"
coverage = "^7.6.2"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"
testcontainers = "^4.8.2"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
